refactored for new backend

+++ app/src/main/java/com/example/investforecast/data/nw/model/StockPrices.kt
@@ -14,7 +14,7 @@ data class StockPrices(
     @SerializedName("close")
     val close: Double,
     @SerializedName("volume")
-    val volume: Long,
+    val volume: Int,
     @SerializedName("date")
     val date: Date,
 ) {
